requirements.txt
boto3
botocore
python-dotenv
tomli; python_version < "3.11"
invoke
requests
pre-commit
